move STreeIndex in Archives
