+++ src/test/java/zArchive/sj/simpleDB/treeIndexing/TreeMapTests.java
@@ -1,4 +1,4 @@
-package db.sTreeIndex;
+package zArchive.sj.simpleDB.treeIndexing;
 
 import java.util.Map.Entry;
 import java.util.SortedMap;
