+++ src/test/java/zArchive/sj/simpleDB/treeIndexing/package-info.java
@@ -5,4 +5,4 @@
  * @author coucou cé moua
  *
  */
-package db.sTreeIndex;
+package zArchive.sj.simpleDB.treeIndexing;
